backend/pyproject.toml
[project]
name = "lua-backend"
version = "2.0.0"
description = "Sistema de Gerenciamento de Joalheria com Assistente Virtual LUA"
authors = [
    {name = "LUA Team", email = "lua@joalheria.com"}
]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "flask-sqlalchemy>=3.1.1",
    "sqlalchemy>=2.0.21",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "bcrypt>=4.0.1",
    "pyjwt>=2.8.0",
    "python-dotenv>=1.0.0",
    "pydub>=0.25.1",
    "numpy>=1.24.3",
    "scipy>=1.11.3",
    "librosa>=0.10.1",
    "soundfile>=0.12.1",
    "transformers>=4.35.2",
    "openai>=1.3.5",
    "ollama>=0.1.7",
    "redis>=5.0.1",
    "cachetools>=5.3.2",
    "requests>=2.31.0",
    "pillow>=10.0.1",
    "python-dateutil>=2.8.2",
    "pydantic>=2.5.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
]

monitoring = [
    "prometheus-client>=0.19.0",
    "sentry-sdk[flask]>=1.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "black>=23.12.0",
    "flake8>=6.1.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"